* change song_paths to a set of paths and raise an exception of duplicate tracks attempted to be added. * if a crates file already exists, do not overwrite it unless a special flag is set. This allows to incfrementally build up crates so that parent crates with content are not wiped in the process of creating the child crates. Add test for this functionality. * black format

+++ src/pyserato/crate.py
@@ -4,6 +4,7 @@ from typing import Iterator, Optional
 from typing_extensions import Self
 
 from pyserato import util
+from pyserato.util import DuplicateTrackError
 
 DEFAULT_SERATO_FOLDER = Path(os.path.expanduser("~/Music/_Serato_"))
 
@@ -11,26 +12,35 @@ DEFAULT_SERATO_FOLDER = Path(os.path.expanduser("~/Music/_Serato_"))
 class Crate:
     def __init__(self, name: str, children: Optional[list[Self]] = None):
         self._children = children if children else []
-        self.filename = util.sanitize_filename(name)
-        self._song_paths: list[Path] = []
+        self.name = util.sanitize_filename(name)
+        self._song_paths: set[Path] = set()
 
     @property
     def children(self) -> list[Self]:
         return self._children
 
     @property
-    def song_paths(self) -> list[Path]:
+    def song_paths(self) -> set[Path]:
         return self._song_paths
 
-    def add_song(self, song_path: str, user_root: str = ''):
+    def add_song(self, song_path: str, user_root: str = ""):
         """
+        Adds a unique song path to the crate. Raises DuplicateTrackError if song path is already present in the Crate.
         :param song_path:
         :param user_root: Support adding an arbitrary root to the songs.
         This is useful when run in a docker container and the path needs to refer to one on the host.
         :return:
         """
         resolved = Path(user_root + song_path).resolve()
-        self._song_paths.append(resolved)
+        if resolved in self._song_paths:
+            raise DuplicateTrackError(f"path {resolved} is already in the crate {self.name}")
+        self._song_paths.add(resolved)
+
+    def __str__(self):
+        return f"Crate<{self.name}>"
+
+    def __repr__(self):
+        return f"Crate<{self.name}>"
 
 
 class Builder:
@@ -44,12 +54,12 @@ class Builder:
         crates = [(root, path)]
         while crates:
             crate, path = crates.pop()
-            path += f'{crate.filename}%%'
+            path += f"{crate.name}%%"
             children = crate.children
             if children:
                 for child in children:
                     crates.append((child, path))
-            yield crate, path.rstrip('%%') + '.crate'
+            yield crate, path.rstrip("%%") + ".crate"
 
     def _build_crate_filepath(self, crate: Crate, serato_folder: Path) -> Iterator[tuple[Crate, Path]]:
         subcrate_folder = serato_folder / "SubCrates"
@@ -58,9 +68,7 @@ class Builder:
 
     @staticmethod
     def _build_save_buffer(crate: Crate) -> bytes:
-        header = (
-            "vrsn   8 1 . 0 / S e r a t o   S c r a t c h L i v e   C r a t e"
-        ).replace(" ", "\0")
+        header = ("vrsn   8 1 . 0 / S e r a t o   S c r a t c h L i v e   C r a t e").replace(" ", "\0")
 
         playlist_section = bytes()
         if crate.song_paths:
@@ -78,7 +86,14 @@ class Builder:
         contents = header.encode() + playlist_section
         return contents
 
-    def save(self, root: Crate, save_path: Path = DEFAULT_SERATO_FOLDER):
+    def save(
+        self,
+        root: Crate,
+        save_path: Path = DEFAULT_SERATO_FOLDER,
+        overwrite: bool = False,
+    ):
         for crate, filepath in self._build_crate_filepath(root, save_path):
+            if filepath.exists() and overwrite is False:
+                continue
             buffer = self._build_save_buffer(crate)
             filepath.write_bytes(buffer)
